+++ lib/varlen_packet_framer_impl.cc
@@ -19,7 +19,7 @@
 #include "varlen_packet_framer_impl.h"
 
 extern "C" {
-#include <fec.h>
+#include "libfec/fec.h"
 #include "golay24.h"
 }