+++ lib/varlen_packet_tagger_impl.cc
@@ -18,7 +18,7 @@
 #include "varlen_packet_tagger_impl.h"
 
 extern "C" {
-#include <fec.h>
+#include "libfec/fec.h"
 #include "golay24.h"
 }